pyproject.toml
[project]
name = "cv-models-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "plotly>=6.3.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
